+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ID-Columns for Users and groups in admin-area
 - Node-ID in label-editor
 - IDs of Filesystems (Datasources) in UI
+- Button in Pipeline-Overview to pause/unpause
 ### Changed
 - Commented out unused variable "datastoreList" for datasets